requirements_new.txt
absl-py==2.1.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
arrow==1.3.0
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
einops==0.8.1
ffmpeg-python==0.2.0
fsspec==2024.10.0
grpcio==1.68.0
httpx==0.27.2
imageio==2.37.0
imageio-ffmpeg==0.6.0
ipykernel==6.29.5
ipython==8.12.3
jax==0.5.0
jaxlib==0.5.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jupyter==1.1.1
jupyterlab==4.3.5
kiwisolver==1.4.7
librosa==0.10.2.post1
matplotlib==3.9.2
mediapipe==0.10.21
numpy==1.26.4
opencv-python==4.11.0.86
pandas==2.2.3
pillow==10.4.0
psutil==6.1.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
